리팩토링(WinningLottoView) : WinningLotto 협력객체 전용 뷰 클래스 추출
기존 메인객체 전체 뷰로부터 협력객체 전용 뷰 로직 클래스를 추출함

WinningLottoView
LottoGameView
